fix: change html-to-image way to screenshot with puppeteer
